Remove WEBSOCKET_UNIFIED_SYSTEM.md and refactor WebSocketManager to improve structure and maintain backward compatibility. Update WebSocket hooks to support flexible authentication and automatic connection handling across various components.
