+++ src/components/websocket/index.ts
@@ -1,0 +1,9 @@
+/**
+ * WebSocket Components
+ * 
+ * This directory contains components related to WebSocket functionality.
+ * The main component is WebSocketManager which is responsible for
+ * managing a single WebSocket connection for the entire application.
+ */
+
+export { default as WebSocketManager } from './WebSocketManager';
